backend/pyproject.toml
[project]
name = "solar-soiling-optimizer"
version = "0.1.0"
description = "Multi-agent Solar Soiling & Yield Optimizer API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "httpx>=0.28.0",
  "pandas>=2.3.0",
  "pydantic-settings>=2.10.0",
  "openai-agents>=0.14.2",
]

[dependency-groups]
dev = [
  "pytest>=8.4.0",
  "ruff>=0.12.0",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["."]
